move emulator to internal package
