+++ integration/internal/emulator/errors.go
