+++ integration/internal/emulator/tests/attachments_test.go
@@ -21,26 +21,30 @@ package tests
 import (
 	"testing"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/onflow/flow-go/integration/emulator"
+	"github.com/onflow/flow-go/integration/internal/emulator"
 )
 
-func TestRuntimeLogs(t *testing.T) {
+func TestAttachments(t *testing.T) {
 
 	t.Parallel()
 
 	b, err := emulator.New()
 	require.NoError(t, err)
 
-	script := []byte(`
-		access(all) fun main() {
-			log("elephant ears")
-		}
-	`)
+	script := `
+		access(all) resource R {}
 
-	result, err := b.ExecuteScript(script, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, []string{`"elephant ears"`}, result.Logs)
+		access(all) attachment A for R {}
+
+		access(all) fun main() {
+			let r <- create R()
+			r[A]
+			destroy r
+		}
+	`
+
+	_, err = b.ExecuteScript([]byte(script), nil)
+	require.NoError(t, err)
 }
