+++ integration/internal/emulator/tests/capcons_test.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/onflow/flow-go/integration/emulator"
+	"github.com/onflow/flow-go/integration/internal/emulator"
 )
 
 func TestCapabilityControllers(t *testing.T) {
