+++ integration/internal/emulator/tests/vm_test.go
@@ -27,7 +27,7 @@ import (
 	"github.com/onflow/flow-go-sdk/test"
 
 	"github.com/onflow/flow-go/fvm"
-	"github.com/onflow/flow-go/integration/emulator"
+	"github.com/onflow/flow-go/integration/internal/emulator"
 	flowgo "github.com/onflow/flow-go/model/flow"
 )
